requirements.txt
Django
whitenoise
psycopg2
dj-database-url
gunicorn
django-extensions
djangorestframework
jwt
djangorestframework-simplejwt
heroku